Test-ClientSettingsConfiguration
https://social.technet.microsoft.com/Forums/en-US/c35d1076-0c03-4fd0-b923-683e066721cd/remediation-script-not-signed?forum=ConfigMgrCompliance

configuration baselines fail, because "script is not signed", even when client settings are set as bypass. -> the reason is that sometimes task-sequences leave some random client settings behind and fails to execute.

highly untested, i have only 1 client policy that is deployed, dunno how it reacts if two or more are applied.