Remove `-Xml` param from `Out-Logfile`
The parameter's not used anywhere so just causes `$Xml` to be shadowed,
and since `$Xml` did not have a default, it's shadowed by `$null`,
resulting in strange behaviour from anything within `Out-Logfile`'s
scope that expects `$Xml` to be defined and populated with a config
file if `Out-Logfile` is called without `-Xml $xml`.

Decided to remove the parameter rather than default to `$script:xml` as
this seemed the clearer option.