Update Function TestCCMCertificateError
We had a rare issue when we had to rebuild our SCCM server from a backup. Clients which were not connected to our internal network/AD site, were not able to renew their Management Point certificate to the new server and therefore lost the communication to the new Management Point. The solution was to run ccmsetup.exe with all parameters + "RESETKEYINFORMATION=TRUE". This fixed our clients in Azure AD and in our other AD forests.
I have added this error and fix to the function TestCCMCertificateError. Also I have renamed some variables, because I had to add some more.